Remove test code from the production code
