+++ package.json
@@ -7,7 +7,7 @@
   "main": "dist/index.js",
   "scripts": {
     "build": "ncc build index.ts",
-    "test": "ts-node index.ts test"
+    "test": "ts-node ./test/test.ts test"
   },
   "dependencies": {
     "@actions/core": "~1.2.7",
